Fixed a problem where shared names in playlist ids were affecting different items because of how ImGui works.
I suppose I addressed an issue where cvars were sometimes pointing to memory that got deallocated... Because it's not explained that ****maybe**** a new cvar is created when a new value is assigned to one? As if they're immutable? But that's not explained, that's just my speculation. Because...

Yeah, no, it's not in documentation: https://wiki.bakkesplugins.com/code_snippets/how_cvars_work/